Move tests dir to root dir
